refactor: demote ComparisonHelper to functions in unordered_list_comparison.py
Per issue #134's helper-consolidation plan (item: comparison_helper -
"Demote or inline; investigate during the PR").

Investigated concrete usage as the issue asked: ComparisonHelper's 3
methods (compare_unordered_lists, unordered_list_metrics, compare_field_raw)
were already 100% @staticmethod with no instance state, and all 3 are
genuinely used from 2 different caller files (structured_model.py,
structured_list_comparator.py) - no dead code here, unlike several earlier
items in this series. That rules out "inline into a single caller"; this
converts them to plain module-level functions instead, same treatment as
result_helper/metrics_helper/rich_value_helper.

Checked for the naming-collision class of bug that hit the rich_value_helper
PR: structured_model.py has an instance method also named compare_field_raw
that delegates to the module-level function of the same name. Confirmed
this is safe - a bare name inside a method body resolves to module/global
scope, not to sibling methods on the same class, so there's no shadowing
(different from the rich_value_helper case, where the collision was a
same-function local parameter, which *does* shadow). Verified with the full
test suite regardless.

No behavior change: full suite passes unchanged, 1541/1541, 2 pre-existing
unrelated skips. No new tests - pure demotion, no new formula or invariant
to pin.

Updates the three usual docs.

Also: investigated field_comparison_helper.py and field_comparison_collector.py
(the other two items marked "Investigate" in the issue) and am deliberately
NOT changing them. Both are built on ComparisonHelperBase, an ABC using the
template-method pattern, which is also the base class for NonMatchesHelper.
The issue itself defers comparison_helper_base.py to a separate "non-match
triplet PRD" I don't have visibility into ("may dissolve as part of that
work"). Restructuring field_comparison_helper.py or field_comparison_collector.py
now risks conflicting with that already-planned, separate effort - flagging
this in the tracker rather than guessing.

Contributes to #134 (9 of 13 helpers; 3 remain not attempted by design -
field_helper.py deferred due to a conflicting open PR #183, and
field_comparison_helper.py/field_comparison_collector.py deferred due to
the ComparisonHelperBase entanglement above).
