+++ src/stickler/structured_object_evaluator/README.md
@@ -370,7 +370,7 @@ The Structured Object Evaluator has been optimized to eliminate redundant Hungar
 - `hungarian_helper.py`: Added unified `get_complete_matching_info()` method
 - `structured_model.py`: Updated to use single Hungarian matching call
 - `structured_list_comparator.py`: Optimized list comparison logic
-- `comparison_helper.py`: Updated helper methods
+- `unordered_list_comparison.py` (then `comparison_helper.py`): Updated helper methods
 - `evaluator_format_helper.py`: Updated evaluator formatting
 
 **Testing**: All 374 existing tests pass, ensuring no regression in functionality.
