+++ src/stickler/structured_object_evaluator/models/structured_list_comparator.py
@@ -14,9 +14,9 @@ Current Behavior Preserved (including bugs):
 from typing import TYPE_CHECKING, Any, Dict, List
 
 from .comparable_field import ComparableField
-from .comparison_helper import ComparisonHelper
 from .hungarian_helper import HungarianHelper
 from .metrics_helper import MetricsHelper
+from .unordered_list_comparison import unordered_list_metrics
 
 if TYPE_CHECKING:
     from .structured_model import StructuredModel
@@ -209,7 +209,7 @@ class StructuredListComparator:
                 0.01  # Almost everything that's not 0.0 should be TP
             )
         
-        match_result = ComparisonHelper.unordered_list_metrics(
+        match_result = unordered_list_metrics(
             threshold_corrected_pairs, gt_list, pred_list, classification_threshold
         )
